moved sortable class helpers to separate file
