+++ src/app/arrange/sortable/Sortable.js
@@ -37,6 +37,27 @@ import {
 	clone,
 	expando,
 } from "./utils.js";
+import {
+	_cancelNextTick,
+	_checkOutsideTargetEl,
+	_detectDirection,
+	_detectNearestEmptySortable,
+	_disableDraggable,
+	_dragElInRowColumn,
+	_generateId,
+	_getSwapDirection,
+	_ghostIsFirst,
+	_ghostIsLast,
+	_globalDragOver,
+	_hideGhostForTarget,
+	_nextTick,
+	_prepareGroup,
+	_saveInputCheckedState,
+	_unhideGhostForTarget,
+	checkCssPointerEventSupport,
+	nearestEmptyInsertDetectEvent,
+	onMove,
+} from "./helpers/sortable.js";
 
 let dragEl,
 	parentEl,
@@ -77,180 +98,12 @@ let dragEl,
 const documentExists = typeof document !== "undefined";
 
 const PositionGhostAbsolutely = IOS;
-const CSSFloatProperty = Edge || IE11OrLess ? "cssFloat" : "float";
 // This will not pass for IE9, because IE9 DnD only works on anchors
 const supportDraggable =
 	documentExists &&
 	!ChromeForAndroid &&
 	!IOS &&
 	"draggable" in document.createElement("div");
-
-const supportCssPointerEvents = (() => {
-	if (!documentExists) return;
-	// false when <= IE11
-	if (IE11OrLess) {
-		return false;
-	}
-	let el = document.createElement("x");
-	el.style.cssText = "pointer-events:auto";
-	return el.style.pointerEvents === "auto";
-})();
-
-const _detectDirection = (el, options) => {
-	let elCSS = css(el),
-		elWidth =
-			parseInt(elCSS.width) -
-			parseInt(elCSS.paddingLeft) -
-			parseInt(elCSS.paddingRight) -
-			parseInt(elCSS.borderLeftWidth) -
-			parseInt(elCSS.borderRightWidth),
-		child1 = getChild(el, 0, options),
-		child2 = getChild(el, 1, options),
-		firstChildCSS = child1 && css(child1),
-		secondChildCSS = child2 && css(child2),
-		firstChildWidth =
-			firstChildCSS &&
-			parseInt(firstChildCSS.marginLeft) +
-				parseInt(firstChildCSS.marginRight) +
-				getRect(child1).width,
-		secondChildWidth =
-			secondChildCSS &&
-			parseInt(secondChildCSS.marginLeft) +
-				parseInt(secondChildCSS.marginRight) +
-				getRect(child2).width;
-
-	if (elCSS.display === "flex") {
-		return elCSS.flexDirection === "column" ||
-			elCSS.flexDirection === "column-reverse"
-			? "vertical"
-			: "horizontal";
-	}
-
-	if (elCSS.display === "grid") {
-		return elCSS.gridTemplateColumns.split(" ").length <= 1
-			? "vertical"
-			: "horizontal";
-	}
-
-	if (child1 && firstChildCSS.float && firstChildCSS.float !== "none") {
-		let touchingSideChild2 = firstChildCSS.float === "left" ? "left" : "right";
-
-		return child2 &&
-			(secondChildCSS.clear === "both" ||
-				secondChildCSS.clear === touchingSideChild2)
-			? "vertical"
-			: "horizontal";
-	}
-
-	return child1 &&
-		(firstChildCSS.display === "block" ||
-			firstChildCSS.display === "flex" ||
-			firstChildCSS.display === "table" ||
-			firstChildCSS.display === "grid" ||
-			(firstChildWidth >= elWidth && elCSS[CSSFloatProperty] === "none") ||
-			(child2 &&
-				elCSS[CSSFloatProperty] === "none" &&
-				firstChildWidth + secondChildWidth > elWidth))
-		? "vertical"
-		: "horizontal";
-};
-
-const _dragElInRowColumn = (dragRect, targetRect, vertical) => {
-	let dragElS1Opp = vertical ? dragRect.left : dragRect.top,
-		dragElS2Opp = vertical ? dragRect.right : dragRect.bottom,
-		dragElOppLength = vertical ? dragRect.width : dragRect.height,
-		targetS1Opp = vertical ? targetRect.left : targetRect.top,
-		targetS2Opp = vertical ? targetRect.right : targetRect.bottom,
-		targetOppLength = vertical ? targetRect.width : targetRect.height;
-
-	return (
-		dragElS1Opp === targetS1Opp ||
-		dragElS2Opp === targetS2Opp ||
-		dragElS1Opp + dragElOppLength / 2 === targetS1Opp + targetOppLength / 2
-	);
-};
-
-/**
- * Detects first nearest empty sortable to X and Y position using emptyInsertThreshold.
- * @param  {Number} x      X position
- * @param  {Number} y      Y position
- * @return {HTMLElement}   Element of the first found nearest Sortable
- */
-const _detectNearestEmptySortable = (x, y) => {
-	let ret;
-	sortables.some((sortable) => {
-		const threshold = sortable[expando].options.emptyInsertThreshold;
-		if (!threshold || lastChild(sortable)) return;
-
-		const rect = getRect(sortable),
-			insideHorizontally =
-				x >= rect.left - threshold && x <= rect.right + threshold,
-			insideVertically =
-				y >= rect.top - threshold && y <= rect.bottom + threshold;
-
-		if (insideHorizontally && insideVertically) {
-			return (ret = sortable);
-		}
-	});
-	return ret;
-};
-
-const _prepareGroup = (options) => {
-	function toFn(value, pull) {
-		return function (to, from, dragEl, evt) {
-			let sameGroup =
-				to.options.group.name &&
-				from.options.group.name &&
-				to.options.group.name === from.options.group.name;
-
-			if (value == null && (pull || sameGroup)) {
-				// Default pull value
-				// Default pull and put value if same group
-				return true;
-			} else if (value == null || value === false) {
-				return false;
-			} else if (pull && value === "clone") {
-				return value;
-			} else if (typeof value === "function") {
-				return toFn(value(to, from, dragEl, evt), pull)(to, from, dragEl, evt);
-			} else {
-				let otherGroup = (pull ? to : from).options.group.name;
-
-				return (
-					value === true ||
-					(typeof value === "string" && value === otherGroup) ||
-					(value.join && value.indexOf(otherGroup) > -1)
-				);
-			}
-		};
-	}
-
-	let group = {};
-	let originalGroup = options.group;
-
-	if (!originalGroup || typeof originalGroup != "object") {
-		originalGroup = { name: originalGroup };
-	}
-
-	group.name = originalGroup.name;
-	group.checkPull = toFn(originalGroup.pull, true);
-	group.checkPut = toFn(originalGroup.put);
-	group.revertClone = originalGroup.revertClone;
-
-	options.group = group;
-};
-
-const _hideGhostForTarget = () => {
-	if (!supportCssPointerEvents && ghostEl) {
-		css(ghostEl, "display", "none");
-	}
-};
-
-const _unhideGhostForTarget = () => {
-	if (!supportCssPointerEvents && ghostEl) {
-		css(ghostEl, "display", "");
-	}
-};
 
 // #1184 fix - Prevent click event on fallback if dragged but item not changed position
 if (documentExists && !ChromeForAndroid) {
@@ -269,32 +122,14 @@ if (documentExists && !ChromeForAndroid) {
 	);
 }
 
-let nearestEmptyInsertDetectEvent = function (evt) {
-	if (dragEl) {
-		evt = evt.touches ? evt.touches[0] : evt;
-		let nearest = _detectNearestEmptySortable(evt.clientX, evt.clientY);
-
-		if (nearest) {
-			// Create imitation event
-			let event = {};
-			for (let i in evt) {
-				if (evt.hasOwnProperty(i)) {
-					event[i] = evt[i];
-				}
-			}
-			event.target = event.rootEl = nearest;
-			event.preventDefault = void 0;
-			event.stopPropagation = void 0;
-			nearest[expando]._onDragOver(event);
+// Fixed #973:
+if (documentExists) {
+	on(document, "touchmove", function (evt) {
+		if ((Sortable.active || awaitingDragStarted) && evt.cancelable) {
+			evt.preventDefault();
 		}
-	}
-};
-
-let _checkOutsideTargetEl = function (evt) {
-	if (dragEl) {
-		dragEl.parentNode[expando]._isOutsideThisEl(evt.target);
-	}
-};
+	});
+}
 
 /**
  * @class  Sortable
@@ -458,7 +293,7 @@ class Sortable {
 				target,
 			filter = options.filter;
 
-		_saveInputCheckedState(el);
+		_saveInputCheckedState(el, savedInputChecked);
 
 		// Don't trigger start event when an element is been dragged, otherwise the evt.oldindex always wrong when set option.group.
 		if (dragEl) {
@@ -592,9 +427,12 @@ class Sortable {
 				find(dragEl, criteria.trim(), _disableDraggable);
 			});
 
-			on(ownerDocument, "dragover", nearestEmptyInsertDetectEvent);
-			on(ownerDocument, "mousemove", nearestEmptyInsertDetectEvent);
-			on(ownerDocument, "touchmove", nearestEmptyInsertDetectEvent);
+			const onCB = (evt) =>
+				nearestEmptyInsertDetectEvent(evt, dragEl, sortables);
+
+			on(ownerDocument, "dragover", onCB);
+			on(ownerDocument, "mousemove", onCB);
+			on(ownerDocument, "touchmove", onCB);
 
 			on(ownerDocument, "mouseup", _this._onDrop.bind(_this));
 			on(ownerDocument, "touchend", _this._onDrop.bind(_this));
@@ -701,7 +539,7 @@ class Sortable {
 		awaitingDragStarted = false;
 		if (rootEl && dragEl) {
 			if (this.nativeDraggable) {
-				on(document, "dragover", _checkOutsideTargetEl);
+				on(document, "dragover", (evt) => _checkOutsideTargetEl(evt, dragEl));
 			}
 			let options = this.options;
 
@@ -722,7 +560,7 @@ class Sortable {
 			this._lastX = touchEvt.clientX;
 			this._lastY = touchEvt.clientY;
 
-			_hideGhostForTarget();
+			_hideGhostForTarget(ghostEl, documentExists);
 
 			let target = document.elementFromPoint(
 				touchEvt.clientX,
@@ -765,7 +603,7 @@ class Sortable {
 				);
 			}
 
-			_unhideGhostForTarget();
+			_unhideGhostForTarget(ghostEl);
 		}
 	}
 
@@ -1063,7 +901,7 @@ class Sortable {
 				dragEl.parentNode[expando]._isOutsideThisEl(evt.target);
 
 				// Do not detect for empty insert if already inserted
-				!insertion && nearestEmptyInsertDetectEvent(evt);
+				!insertion && nearestEmptyInsertDetectEvent(evt, dragEl, sortables);
 			}
 
 			!options.dragoverBubble && evt.stopPropagation && evt.stopPropagation();
@@ -1239,7 +1077,11 @@ class Sortable {
 						? options.swapThreshold
 						: options.invertedSwapThreshold,
 					isCircumstantialInvert,
-					lastTarget === target
+					lastTarget === target,
+					dragEl,
+					targetMoveDistance,
+					pastFirstInvertThresh,
+					lastDirection
 				);
 
 				let sibling;
@@ -1287,7 +1129,7 @@ class Sortable {
 					}
 
 					_silent = true;
-					setTimeout(_unsilent, 30);
+					setTimeout(() => (_silent = false), 30);
 
 					capture();
 
@@ -1332,12 +1174,15 @@ class Sortable {
 	}
 
 	_offMoveEvents() {
+		const offCB = (evt) =>
+			nearestEmptyInsertDetectEvent(evt, dragEl, sortables);
+
 		off(document, "mousemove", this._onTouchMove.bind(this));
 		off(document, "touchmove", this._onTouchMove.bind(this));
 		off(document, "pointermove", this._onTouchMove.bind(this));
-		off(document, "dragover", nearestEmptyInsertDetectEvent);
-		off(document, "mousemove", nearestEmptyInsertDetectEvent);
-		off(document, "touchmove", nearestEmptyInsertDetectEvent);
+		off(document, "dragover", offCB);
+		off(document, "mousemove", offCB);
+		off(document, "touchmove", offCB);
 	}
 
 	_offUpEvents() {
@@ -1668,219 +1513,6 @@ class Sortable {
 			cloneHidden = false;
 		}
 	}
-}
-
-function _globalDragOver(/**Event*/ evt) {
-	if (evt.dataTransfer) {
-		evt.dataTransfer.dropEffect = "move";
-	}
-	evt.cancelable && evt.preventDefault();
-}
-
-function onMove(
-	fromEl,
-	toEl,
-	dragEl,
-	dragRect,
-	targetEl,
-	targetRect,
-	originalEvent,
-	willInsertAfter
-) {
-	let evt,
-		sortable = fromEl[expando],
-		onMoveFn = sortable.options.onMove,
-		retVal;
-	// Support for new CustomEvent feature
-	if (window.CustomEvent && !IE11OrLess && !Edge) {
-		evt = new CustomEvent("move", {
-			bubbles: true,
-			cancelable: true,
-		});
-	} else {
-		evt = document.createEvent("Event");
-		evt.initEvent("move", true, true);
-	}
-
-	evt.to = toEl;
-	evt.from = fromEl;
-	evt.dragged = dragEl;
-	evt.draggedRect = dragRect;
-	evt.related = targetEl || toEl;
-	evt.relatedRect = targetRect || getRect(toEl);
-	evt.willInsertAfter = willInsertAfter;
-
-	evt.originalEvent = originalEvent;
-
-	fromEl.dispatchEvent(evt);
-
-	if (onMoveFn) {
-		retVal = onMoveFn.call(sortable, evt, originalEvent);
-	}
-
-	return retVal;
-}
-
-function _disableDraggable(el) {
-	el.draggable = false;
-}
-
-function _unsilent() {
-	_silent = false;
-}
-
-function _ghostIsFirst(evt, vertical, sortable) {
-	let rect = getRect(getChild(sortable.el, 0, sortable.options, true));
-	const spacer = 10;
-
-	return vertical
-		? evt.clientX < rect.left - spacer ||
-				(evt.clientY < rect.top && evt.clientX < rect.right)
-		: evt.clientY < rect.top - spacer ||
-				(evt.clientY < rect.bottom && evt.clientX < rect.left);
-}
-
-function _ghostIsLast(evt, vertical, sortable) {
-	let rect = getRect(lastChild(sortable.el, sortable.options.draggable));
-	const spacer = 10;
-
-	return vertical
-		? evt.clientX > rect.right + spacer ||
-				(evt.clientX <= rect.right &&
-					evt.clientY > rect.bottom &&
-					evt.clientX >= rect.left)
-		: (evt.clientX > rect.right && evt.clientY > rect.top) ||
-				(evt.clientX <= rect.right && evt.clientY > rect.bottom + spacer);
-}
-
-function _getSwapDirection(
-	evt,
-	target,
-	targetRect,
-	vertical,
-	swapThreshold,
-	invertedSwapThreshold,
-	invertSwap,
-	isLastTarget
-) {
-	let mouseOnAxis = vertical ? evt.clientY : evt.clientX,
-		targetLength = vertical ? targetRect.height : targetRect.width,
-		targetS1 = vertical ? targetRect.top : targetRect.left,
-		targetS2 = vertical ? targetRect.bottom : targetRect.right,
-		invert = false;
-
-	if (!invertSwap) {
-		// Never invert or create dragEl shadow when target movemenet causes mouse to move past the end of regular swapThreshold
-		if (isLastTarget && targetMoveDistance < targetLength * swapThreshold) {
-			// multiplied only by swapThreshold because mouse will already be inside target by (1 - threshold) * targetLength / 2
-			// check if past first invert threshold on side opposite of lastDirection
-			if (
-				!pastFirstInvertThresh &&
-				(lastDirection === 1
-					? mouseOnAxis > targetS1 + (targetLength * invertedSwapThreshold) / 2
-					: mouseOnAxis < targetS2 - (targetLength * invertedSwapThreshold) / 2)
-			) {
-				// past first invert threshold, do not restrict inverted threshold to dragEl shadow
-				pastFirstInvertThresh = true;
-			}
-
-			if (!pastFirstInvertThresh) {
-				// dragEl shadow (target move distance shadow)
-				if (
-					lastDirection === 1
-						? mouseOnAxis < targetS1 + targetMoveDistance // over dragEl shadow
-						: mouseOnAxis > targetS2 - targetMoveDistance
-				) {
-					return -lastDirection;
-				}
-			} else {
-				invert = true;
-			}
-		} else {
-			// Regular
-			if (
-				mouseOnAxis > targetS1 + (targetLength * (1 - swapThreshold)) / 2 &&
-				mouseOnAxis < targetS2 - (targetLength * (1 - swapThreshold)) / 2
-			) {
-				return _getInsertDirection(target);
-			}
-		}
-	}
-
-	invert = invert || invertSwap;
-
-	if (invert) {
-		// Invert of regular
-		if (
-			mouseOnAxis < targetS1 + (targetLength * invertedSwapThreshold) / 2 ||
-			mouseOnAxis > targetS2 - (targetLength * invertedSwapThreshold) / 2
-		) {
-			return mouseOnAxis > targetS1 + targetLength / 2 ? 1 : -1;
-		}
-	}
-
-	return 0;
-}
-
-/**
- * Gets the direction dragEl must be swapped relative to target in order to make it
- * seem that dragEl has been "inserted" into that element's position
- * @param  {HTMLElement} target       The target whose position dragEl is being inserted at
- * @return {Number}                   Direction dragEl must be swapped
- */
-function _getInsertDirection(target) {
-	if (index(dragEl) < index(target)) {
-		return 1;
-	} else {
-		return -1;
-	}
-}
-
-/**
- * Generate id
- * @param   {HTMLElement} el
- * @returns {String}
- * @private
- */
-function _generateId(el) {
-	let str = el.tagName + el.className + el.src + el.href + el.textContent,
-		i = str.length,
-		sum = 0;
-
-	while (i--) {
-		sum += str.charCodeAt(i);
-	}
-
-	return sum.toString(36);
-}
-
-function _saveInputCheckedState(root) {
-	savedInputChecked.length = 0;
-
-	let inputs = root.getElementsByTagName("input");
-	let idx = inputs.length;
-
-	while (idx--) {
-		let el = inputs[idx];
-		el.checked && savedInputChecked.push(el);
-	}
-}
-
-function _nextTick(fn) {
-	return setTimeout(fn, 0);
-}
-
-function _cancelNextTick(id) {
-	return clearTimeout(id);
-}
-
-// Fixed #973:
-if (documentExists) {
-	on(document, "touchmove", function (evt) {
-		if ((Sortable.active || awaitingDragStarted) && evt.cancelable) {
-			evt.preventDefault();
-		}
-	});
 }
 
 // Export utils
